backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt==4.0.1